requirements.txt
streamlit
openai-whisper
pytubefix
numpy
pandas
torch
torchvision
torchaudio
pytest
python-dotenv
requests
huggingface_hub